Add files via upload
The information is presented in PDF (Portable Document Format), it corresponds to 104 personalized documents, each document consists of at least 4 pages in free format and in Spanish, the review and compliance with the regulations for public information indicated in the article. 113, section I of the Federal Law of Transparency and Access to Public Information, as well as the corresponding scan was carried out by the National Council for the Evaluation of Social Development Policy (CONEVAL), obtained from the site https: //www.coneval .org.mx / quienessomos / InvestigadoresAcademicos / Paginas / Convocatoria_Eleccion_2020.aspx.