Create README.md
The importance of the working document is that it allows analyzing the position of 104 researchers of the Conacyt National System of Researchers (Candidate levels, Level I, Level II and Level III) who present their perspective on the importance of the evaluation and measurement of poverty for the development of social policy and the challenges for the consolidation of evaluation in Mexico, as part of the call dated January 9, 2020 for the process of electing three academic researchers to join the National Council for the Evaluation of the Social Development Policy in the National Council for the Evaluation of Social Development Policy (CONEVAL). The text analysis is carried out by applying a text mining algorithm using the Orange Data Mining 3.26.0 software, which provides fast and timely information, in order to identify thematic groups of the positioning, in order to Identify area of ​​opportunity in such a way that the information serves as the basis for analysis and planning in the different areas of knowledge regarding Social Development in Mexico.